Split complex header file and tests into multiple files
