+++ sycl/include/sycl/ext/oneapi/experimental/complex/complex.hpp
@@ -1,0 +1,16 @@
+//===- complex.hpp --------------------------------------------------------===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#pragma once
+
+#ifdef SYCL_EXT_ONEAPI_COMPLEX
+
+#include "./detail/complex.hpp"
+#include "./detail/complex_math.hpp"
+
+#endif // SYCL_EXT_ONEAPI_COMPLEX
